fix(schedule): move ScheduleLeaseLostError to domain layer
ScheduleLeaseLostError was defined in infrastructure/lease but consumed by
application services, violating the application-must-not-import-infrastructure
boundary (CIS Governance check failed). Move the class to
server/domain/errors/schedule-lease-lost-error.ts; all five consumers (2 app
services, 2 repositories, lease coordinator) now import from the domain barrel.

Package-internal-boundary-audit now passes (0 violations).
