+++ packages/schedule/src/server/domain/errors/index.ts
@@ -1,0 +1,2 @@
+// Schedule errors - business-specific error classes
+export * from './schedule-lease-lost-error';
