+++ packages/schedule/src/server/infrastructure/adapters/powersync/schedule-powersync.repository.ts
@@ -1,6 +1,6 @@
 import type { IScheduleRepository, ScheduleRebuildOutboxDTO } from '../../../domain/repositories/i-schedule-repository';
 import type { CalendarEntry } from '../../../domain/aggregates/calendar-entry';
-import { ScheduleLeaseLostError } from '../../lease/schedule-lease-coordinator';
+import { ScheduleLeaseLostError } from '../../../domain/errors/schedule-lease-lost-error';
 import {
   PowerSyncScheduleMapper,
   type PowerSyncScheduleRow,
